First prototype of more modular and customizable color styles.
Prototype is at top level, overriding the Noteworthy theme. However,
the idea is to propose changes to the theme going forward.

See: assets/css/{main,colors,color_palette}.scss

In this scheme, a user of the theme that wanted to simply change
colors would override the `color_palette.scss` file defining new
colors and mapping them to Noteworthy's two basic colors--called here,
"primary" and "secondary"--with shades.

If the user wanted to change the use of the basic colors in the css
rules (e.g. use the $secondary-darker color for visited links rather
than the theme's use of $secondary), she would override the
`colors.scss` file and change the appropriate variable definitions.

One, or both of these files can be overriden.

This scheme has the following advantages:
1.  The main.scss file does not need to be overriden for color changes.
    This large file is likely to change with the theme's evolution,
    making it hard for the user to keep the override up to date.
2.  The color usage is made more explicit.
3.  If one wants to change the secondary color ("teal"), it avoids
    the hack of simply re-defining "$teal" to some other color that
    is not related to the color teal.
4.  With the mapping of color definitions to usage, one avoids having
    to find all the uses of a color variable, figuring out what the
    usage is, and then making changes if desired in the main css file.

The `main.scss` can still be overriden if needed with or without
overriding the `colors.scss` or `color_palette.scss` files.

This new scheme retains the colors of the Noteworthy theme (the
generated css file is the same).  The body of `main.scss` has the
following changes:
1.  The color definitions have been removed.  They are now imported via
    the `colors.scss` file (which imports the `color_palette.scss`
    file).
2.  The color variables in the style rules have been replaced with
    new variables having names matching their usage (see `colors.scss`).

The first change is small textually and is easy for users to accomodate
if they have already overriden `main.scss`.

The second change would be problematic if users have overriden
`main.scss` and made significant edits; even edits not involving color.
It *might* be possible to use some sort of merge/diff tool, but this has
not yet been explored.  It is exactly this kind of problem--overriding
and customizing a file that is large, complex, and addressing many
concerns--that this color scheme addresses!

Note that if the user has provided a custom css with css overrides, they
are not impacted by this change.